Move logic to create navigation to MainActivity
